fix: refactored the whole code, wrapped client into a struct
