+++ cli.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+// "fmt"
+// "os"
+// "github.com/codescalersinternships/DateTime-HTTP-Client-Rowan/client"
+)
+
+func main() {
+	// client := NewDateTimeClient()
+	// body, _ := client.GetDateTime()
+	// fmt.Printf("%s\n", body)
+	// _, err:= client.GetDateTime(os.Getenv("URL"), "rowann")
+	// fmt.Println(err)
+	// client.GetDateTime("http://localhost:8080/datetime")
+
+}
